Jamesmccomish/update entrypoint

+++ src/utils/Utils.sol
@@ -22,69 +22,69 @@ pragma solidity ^0.8.3;
  * @author Modified from Argent (https://github.com/argentlabs/argent-contracts/blob/develop/contracts/modules/common/Utils.sol)
  */
 library Utils {
-	/**
-	 * @notice Helper method to recover the signer at a given position from a list of concatenated signatures.
-	 * @param _signedHash The signed hash
-	 * @param _signatures The concatenated signatures.
-	 * @param _index The index of the signature to recover.
-	 */
-	function recoverSigner(
-		bytes32 _signedHash,
-		bytes memory _signatures,
-		uint _index
-	) internal pure returns (address) {
-		uint8 v;
-		bytes32 r;
-		bytes32 s;
-		// we jump 32 (0x20) as the first slot of bytes contains the length
-		// we jump 65 (0x41) per signature
-		// for v we load 32 bytes ending with v (the first 31 come from s) then apply a mask
-		// solhint-disable-next-line no-inline-assembly
-		assembly {
-			r := mload(add(_signatures, add(0x20, mul(0x41, _index))))
-			s := mload(add(_signatures, add(0x40, mul(0x41, _index))))
-			v := and(mload(add(_signatures, add(0x41, mul(0x41, _index)))), 0xff)
-		}
-		require(v == 27 || v == 28, 'Utils: bad v value in signature');
+    /**
+     * @notice Helper method to recover the signer at a given position from a list of concatenated signatures.
+     * @param _signedHash The signed hash
+     * @param _signatures The concatenated signatures.
+     * @param _index The index of the signature to recover.
+     */
+    function recoverSigner(bytes32 _signedHash, bytes memory _signatures, uint256 _index)
+        internal
+        pure
+        returns (address)
+    {
+        uint8 v;
+        bytes32 r;
+        bytes32 s;
+        // we jump 32 (0x20) as the first slot of bytes contains the length
+        // we jump 65 (0x41) per signature
+        // for v we load 32 bytes ending with v (the first 31 come from s) then apply a mask
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            r := mload(add(_signatures, add(0x20, mul(0x41, _index))))
+            s := mload(add(_signatures, add(0x40, mul(0x41, _index))))
+            v := and(mload(add(_signatures, add(0x41, mul(0x41, _index)))), 0xff)
+        }
+        require(v == 27 || v == 28, "Utils: bad v value in signature");
 
-		address recoveredAddress = ecrecover(_signedHash, v, r, s);
-		require(recoveredAddress != address(0), 'Utils: ecrecover returned 0');
-		return recoveredAddress;
-	}
+        address recoveredAddress = ecrecover(_signedHash, v, r, s);
+        require(recoveredAddress != address(0), "Utils: ecrecover returned 0");
+        return recoveredAddress;
+    }
 
-	/**
-	 * @notice Helper method to parse data and extract the method signature.
-	 */
-	function functionPrefix(bytes memory _data) internal pure returns (bytes4 prefix) {
-		require(_data.length >= 4, 'Utils: Invalid functionPrefix');
-		// solhint-disable-next-line no-inline-assembly
-		assembly {
-			prefix := mload(add(_data, 0x20))
-		}
-	}
+    /**
+     * @notice Helper method to parse data and extract the method signature.
+     */
+    function functionPrefix(bytes memory _data) internal pure returns (bytes4 prefix) {
+        require(_data.length >= 4, "Utils: Invalid functionPrefix");
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            prefix := mload(add(_data, 0x20))
+        }
+    }
 
-	/**
-	 * @notice Checks if an address is a contract.
-	 * @param _addr The address.
-	 */
-	function isContract(address _addr) internal view returns (bool) {
-		uint32 size;
-		// solhint-disable-next-line no-inline-assembly
-		assembly {
-			size := extcodesize(_addr)
-		}
-		return (size > 0);
-	}
+    /**
+     * @notice Checks if an address is a contract.
+     * @param _addr The address.
+     */
+    function isContract(address _addr) internal view returns (bool) {
+        uint32 size;
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            size := extcodesize(_addr)
+        }
+        return (size > 0);
+    }
 
-	/**
-	 * @notice Returns ceil(a / b).
-	 */
-	function ceil(uint256 a, uint256 b) internal pure returns (uint256) {
-		uint256 c = a / b;
-		if (a % b == 0) {
-			return c;
-		} else {
-			return c + 1;
-		}
-	}
+    /**
+     * @notice Returns ceil(a / b).
+     */
+    function ceil(uint256 a, uint256 b) internal pure returns (uint256) {
+        uint256 c = a / b;
+        if (a % b == 0) {
+            return c;
+        } else {
+            return c + 1;
+        }
+    }
 }
